backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pydantic-settings==2.3.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT[crypto]==2.8.0
celery[redis]==5.4.0
redis==5.0.4
tenacity==8.2.3
aiofiles==23.2.1
orjson==3.10.3
python-dotenv==1.0.1
asyncpg==0.29.0
